Need to be more careful with directive API
In cases where a directive has API surface that would also be a
meaningful DOM attribute (e.g. checked or disabled) we need to erase
that attribute from the rendered element as part of linking the
directive lest it have effects on the runtime behavior of the control.

In this particular case the "disabled" attribute on a div was resulting
in the toggle switch not behaving correctly with regard to its expected
events being delivered.