Adding Handle support to certain directives
Some WinJS widgets have methods e.g. hide/show or open/close that are
not exposed by the current WinJS shim for AngularJS. This patch adds
support for a "handle" attribute on those directives which when set
forces teh shim to put teh wdget instance in the scope as indicated by
teh value of the "handle" attribute.

Signed-off-by: Dylan Borg <borgdylan@hotmail.com>