Fix #48: Class name not on HubSection/PivotItem
The win-hub-section (and win-pivot-item) directive instantiated
WinJS.UI.HubSection on a child of win-hub-section rather than directly on
win-hub-section. Consequently, any attributes that a developer may have
put on win-hub-section (e.g. class) didn't get carried over to the
WinJS.UI.HubSection.

This fixes the issue by instantiating WinJS.UI.HubSection directly on
win-hub-section. Now win-hub-section can't be used as the placeholder
element. To fix that, the placeholder element is created dynamically and
takes the place of win-hub-section as a child of placeholder-holder.